requirements.txt
bitfinex-api-py==3.0.1
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pyee==9.0.4
python-dotenv==1.0.1
requests==2.28.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.19
websockets==11.0.3
